+++ tests/bench/connection/local_lifecycle.py
@@ -27,5 +27,7 @@ def main(router):
         t = mitogen.core.now()
         f = router.local(debug=opts.debug)
         tt = mitogen.core.now()
-        print(x, 1000 * (tt - t))
-    print('%.03f ms' % (1000 * (mitogen.core.now() - t0) / (1.0 + x)))
+
+    t1 = mitogen.core.now()
+    mean = (t1 - t0) / opts.iterations
+    print('++ iterations %d, mean %.03f ms' % (opts.iterations, 1e3 * mean))